requirements-dev.txt
# Development dependencies for code quality
# Install with: pip install -r requirements-dev.txt

# Code formatting
black>=23.0.0

# Linting
flake8>=6.0.0
pylint>=2.17.0

# Type checking
mypy>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# YAML support for config files (optional)
pyyaml>=6.0
